rename chatbot module
