+++ ai-chatbot/proguard-rules.pro
