+++ ai-chatbot/src/main/res/layout/activity_after_login.xml
