+++ ai-chatbot/src/main/res/layout/activity_ai_chatbot.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.likeminds.chatmm.aichatbot.LMChatAIButton
+    <com.likeminds.chatmm.aichatbot.LMChatAIBotButton
         android:id="@+id/btn_ai_chatbot"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
