+++ ai-chatbot/src/main/res/layout/activity_auth.xml
