+++ ai-chatbot/src/main/res/values/colors.xml
