+++ ai-chatbot/src/main/res/values/strings.xml
