+++ settings.gradle
@@ -25,4 +25,4 @@ include ':community-hybrid-chat'
 include ':chatmm'
 include ':community-chat'
 include ':networking-chat'
-include ':community-ai-chatbot'
+include ':ai-chatbot'
